+++ VisualStudio/Hengband/Hengband.vcxproj
@@ -626,6 +626,7 @@
     <ClCompile Include="..\..\src\system\player-type-definition.cpp" />
     <ClCompile Include="..\..\src\system\services\baseitem-monrace-service.cpp" />
     <ClCompile Include="..\..\src\system\terrain\terrain-definition.cpp" />
+    <ClCompile Include="..\..\src\system\terrain\terrain-list.cpp" />
     <ClCompile Include="..\..\src\target\grid-selector.cpp" />
     <ClCompile Include="..\..\src\target\projection-path-calculator.cpp" />
     <ClCompile Include="..\..\src\target\target-describer.cpp" />
@@ -1453,6 +1454,7 @@
     <ClInclude Include="..\..\src\system\services\baseitem-monrace-service.h" />
     <ClInclude Include="..\..\src\system\terrain\terrain-definition.h" />
     <ClInclude Include="..\..\src\system\enums\terrain\terrain-tag.h" />
+    <ClInclude Include="..\..\src\system\terrain\terrain-list.h" />
     <ClInclude Include="..\..\src\target\grid-selector.h" />
     <ClInclude Include="..\..\src\target\projection-path-calculator.h" />
     <ClInclude Include="..\..\src\target\target-describer.h" />
