+++ src/floor/object-allocator.cpp
@@ -16,6 +16,7 @@
 #include "system/monrace/monrace-definition.h"
 #include "system/player-type-definition.h"
 #include "system/terrain/terrain-definition.h"
+#include "system/terrain/terrain-list.h"
 #include "util/bit-flags-calculator.h"
 #include "wizard/wizard-messages.h"
