+++ src/grid/grid.cpp
@@ -50,6 +50,7 @@
 #include "system/player-type-definition.h"
 #include "system/redrawing-flags-updater.h"
 #include "system/terrain/terrain-definition.h"
+#include "system/terrain/terrain-list.h"
 #include "term/gameterm.h"
 #include "term/term-color-types.h"
 #include "timed-effect/timed-effects.h"
