+++ src/info-reader/dungeon-reader.cpp
@@ -10,6 +10,7 @@
 #include "system/dungeon/dungeon-list.h"
 #include "system/monrace/monrace-definition.h"
 #include "system/terrain/terrain-definition.h"
+#include "system/terrain/terrain-list.h"
 #include "util/string-processor.h"
 #include "view/display-messages.h"
 #include <span>
