+++ src/info-reader/feature-reader.cpp
@@ -9,6 +9,7 @@
 #include "main/angband-headers.h"
 #include "room/door-definition.h"
 #include "system/terrain/terrain-definition.h"
+#include "system/terrain/terrain-list.h"
 #include "term/gameterm.h"
 #include "util/bit-flags-calculator.h"
 #include "util/string-processor.h"
