+++ src/info-reader/general-parser.cpp
@@ -18,6 +18,7 @@
 #include "system/floor/floor-info.h"
 #include "system/system-variables.h"
 #include "system/terrain/terrain-definition.h"
+#include "system/terrain/terrain-list.h"
 #include "util/angband-files.h"
 #include "util/string-processor.h"
 #include <string>
