+++ src/io/interpret-pref-file.cpp
@@ -20,6 +20,7 @@
 #include "system/monrace/monrace-list.h"
 #include "system/player-type-definition.h"
 #include "system/terrain/terrain-definition.h"
+#include "system/terrain/terrain-list.h"
 #include "term/gameterm.h"
 #include "util/string-processor.h"
 #include "view/display-messages.h"
