+++ src/main/info-initializer.cpp
@@ -37,6 +37,7 @@
 #include "system/player-type-definition.h"
 #include "system/spell-info-list.h"
 #include "system/terrain/terrain-definition.h"
+#include "system/terrain/terrain-list.h"
 #include "util/angband-files.h"
 #include "util/string-processor.h"
 #include "view/display-messages.h"
