+++ src/player/player-move.cpp
@@ -45,6 +45,7 @@
 #include "system/player-type-definition.h"
 #include "system/redrawing-flags-updater.h"
 #include "system/terrain/terrain-definition.h"
+#include "system/terrain/terrain-list.h"
 #include "target/target-checker.h"
 #include "timed-effect/timed-effects.h"
 #include "util/bit-flags-calculator.h"
