+++ src/room/rooms-city.cpp
@@ -13,6 +13,7 @@
 #include "system/grid-type-definition.h"
 #include "system/player-type-definition.h"
 #include "system/terrain/terrain-definition.h"
+#include "system/terrain/terrain-list.h"
 #include "util/bit-flags-calculator.h"
 #include "wizard/wizard-messages.h"
 #include <algorithm>
