+++ src/room/rooms-trap.cpp
@@ -11,6 +11,7 @@
 #include "system/grid-type-definition.h"
 #include "system/player-type-definition.h"
 #include "system/terrain/terrain-definition.h"
+#include "system/terrain/terrain-list.h"
 #include "wizard/wizard-messages.h"
 
 /*!
