+++ src/view/display-map.cpp
@@ -21,6 +21,7 @@
 #include "system/monster-entity.h"
 #include "system/player-type-definition.h"
 #include "system/terrain/terrain-definition.h"
+#include "system/terrain/terrain-list.h"
 #include "term/term-color-types.h"
 #include "timed-effect/timed-effects.h"
 #include "util/bit-flags-calculator.h"
